% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ct_search}
\alias{ct_search}
\title{Search Clinical Trials using the REST API}
\usage{
ct_search(...)
}
\description{
Filters all clinical trials based upon supplied filter params.
}
\details{
Filter params may be any of the fields in the schema as well as any of the following params...
\itemize{
\item{size: limit the amount of results a supplied amount (default is 10, max is 50)}
\item{from: start the results from a supplied starting point (default is 0)}
\item{include: include only the supplied filter param fields in all results (useful if you want to minimize the payload returned)}
\item{exclude: exclude the supplied filter param fields from all results (useful if you want most of the payload returned with the exception of a few fields)}
\item{fulltext: filter results by examining a variety of text-based fields (including the trial title, description, treatments, etc)}
}
Note that string field values are not case sensitive (must otherwise must match exactly).

For field params which are filtering as ranges (date and long types), please supply _gte or _lte
to the end of the field param (depending on if you are filtering on greater than or equal (gte),
less than or equal (lte), or both): \code{<field_param>_gte=<field_value_from>},
\code{<field_param>_lte=<field_value_to>}

\code{record_verification_date_gte="2016-08-25"}

For field params which are geolocation coordinates (geo_point), please supply the
following to the end of the field param:
\itemize{
\item{\code{_lat} - The latitude in decimal degrees and plus/minus.}
\item{\code{_lon} - The longitude in decimal degrees and plus/minus.}
\item{\code{_dist} - The radius to search within. }
}

Format must be an integer followed by a unit defined as:
\itemize{
\item{mi - miles (for example \code{"2mi"})}
\item{km - kilometer (for example \code{"5km")}}
}

If you are crafting more complicated queries, it might be best to use the POST endpoint of the same name.
}
\examples{
res = ct_search(eligibility.structured.gender="female",
       include="nct_id")
head(res)
res = ct_search(record_verification_date_gte="2016-08-25")
res$count
res = ct_search(sites.org_coordinates_lat=39.1292,
             sites.org_coordinates_lon=-77.2953,
             sites.org_coordinates_dist="100mi")

}
\references{
\url{https://clinicaltrialsapi.cancer.gov/}
}
