% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fields}
\alias{fields}
\title{Return available API fields and data types}
\usage{
fields()
}
\description{
This function simply returns a named character
vector with all available fields for searching
and returning. It is meant to inform building
queries.
}
\references{
\url{https://clinicaltrialsapi.cancer.gov/}
}
